requirements.txt
Django>=4.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=3.13.0
requests>=2.26.0
